Centralize location of serialization/deserialization errors
When implementing user-definable serialization, we were still using our
original forked HighThroughputExecutor. That executor relied on
globus_compute_sdk.errors.SerializationError. At the same time, the
serialization module relied on its own SerializationError. To avoid
mixing their semantics up, I left them in place. Now that we've moved
away from HighThroughputExecutor, we no longer need that differentiation,
so for consistency's sake, this moves the ser/de errors into the SDK's
common errors module.

At the same time, introduce a base class to catch both serialization and
deserialization errors, and ensure that all three error classes are also
ComputeErrors.
