+++ compute_sdk/globus_compute_sdk/errors/__init__.py
@@ -1,6 +1,8 @@
 from .error_types import (
     ComputeError,
+    DeserializationError,
     MaxResultSizeExceeded,
+    SerdeError,
     SerializationError,
     TaskExecutionFailed,
     TaskPending,
@@ -11,7 +13,9 @@ __all__ = (
     "ComputeError",
     "TaskExecutionFailed",
     "MaxResultSizeExceeded",
+    "SerdeError",
     "SerializationError",
+    "DeserializationError",
     "TaskPending",
     "VersionMismatch",
 )
